+ create tests files for any tests namespace
